requirements.txt
langchain
openai
ipykernel
langchain_community
huggingface_hub
python_dotenv
streamlit